Implement ZipDiffValidator and ZipDiffTest

+++ harness/src/main/java/org/jboss/pvt/harness/utils/ResourceUtils.java
@@ -31,16 +31,16 @@ import static org.apache.commons.lang.StringUtils.isNotEmpty;
 /**
  * Created by rnc on 02/10/16.
  */
-public class ResourceUtil
+public class ResourceUtils
 {
-    private static Logger logger = LoggerFactory.getLogger( ResourceUtil.class );
+    private static Logger logger = LoggerFactory.getLogger( ResourceUtils.class );
 
     /**
      *
      * @param zip local zip or http url
      * @return explored dir
      */
-    public static File downloadZips( String zip )
+    public static File downloadZip(String zip )
     {
         File result = null;
         try
@@ -57,7 +57,59 @@ public class ResourceUtil
                     }
                     else
                     {
-                        logger.debug( "Copying URL {} to {}", zip, result );
+                        logger.info( "Copying URL {} to {}", zip, result );
+                        org.apache.commons.io.FileUtils.copyURLToFile( new URL( zip ), result );
+                    }
+                }
+                // Local file path
+                else
+                {
+
+                    result = new File( System.getProperty( "basedir" ) + "/target/" + zip.substring( zip.lastIndexOf( "/" ) + 1 ) );
+                    if ( result.exists() )
+                    {
+                        logger.warn( "Avoiding duplicate download of {} ", zip );
+                    }
+                    else
+                    {
+                        org.apache.commons.io.FileUtils.copyFile( new File( zip ), result );
+                    }
+                }
+                logger.debug( "Create distribution {} ", result );
+            }
+        }
+        catch ( IOException | ZipException e )
+        {
+            logger.error( "Caught ", e );
+            throw new PVTSystemException( "Caught exception processing downloads: ", e );
+        }
+        return result;
+    }
+
+
+    /**
+     *
+     * @param zip local zip or http url
+     * @return explored dir
+     */
+    public static File downloadZipExplored(String zip )
+    {
+        File result = null;
+        try
+        {
+            if ( isNotEmpty( zip ) )
+            {
+                // URL Based form
+                if ( zip.contains( "://" ) )
+                {
+                    result = new File("target/" + zip.substring( zip.lastIndexOf( "/" ) + 1 ) + ".unzipped" );
+                    if ( result.exists())
+                    {
+                        logger.warn( "Avoiding duplicate download of {} ", zip );
+                    }
+                    else
+                    {
+                        logger.info( "Copying URL {} to {}", zip, result );
                         org.apache.commons.io.FileUtils.copyURLToFile( new URL( zip ), result );
                         ZipUtil.explode( result );
                     }
@@ -66,7 +118,7 @@ public class ResourceUtil
                 else
                 {
 
-                    result = new File( System.getProperty( "basedir" ) + "/target/" + zip.substring( zip.lastIndexOf( "/" ) + 1 ) );
+                    result = new File( System.getProperty( "basedir" ) + "/target/" + zip.substring( zip.lastIndexOf( "/" ) + 1 ) + ".unzipped" );
                     if ( result.exists() )
                     {
                         logger.warn( "Avoiding duplicate download of {} ", zip );
@@ -87,4 +139,6 @@ public class ResourceUtil
         }
         return result;
     }
+
+
 }
